Use weather api instead of SWAPI
The Star Wars api only let's you query on id for characters and movies etc.
To have a better representation of the challenge I opted to use a weather api since this would allow for user input to select the city.
FYI: the requested spotify API is unavailable (at this time at least).